requirements.txt
rich==13.3.2
requests==2.28.2
python-dotenv==1.0.0
setuptools-scm==7.1.0
anthropic==0.3.1
fastapi==0.95.1
uvicorn==0.21.1
